Add browser tests
